Show the total number of six nutritions

+++ app/src/main/java/com/terricom/mytype/data/Foodie.kt
@@ -9,11 +9,11 @@ data class Foodie (
     val timestamp: Date ?= null,
     val photo: String ?= "",
     val foods: List<String> ?= null,
-    val water: Long ?= null,
-    val oil: Long ?= null,
-    val vegetable: Long ?= null,
-    val protein: Long ?= null,
-    val fruit: Long ?= null,
-    val carbon: Long ?= null,
+    val water: Float ?= null,
+    val oil: Float ?= null,
+    val vegetable: Float ?= null,
+    val protein: Float ?= null,
+    val fruit: Float ?= null,
+    val carbon: Float ?= null,
     val nutritions: List<String> ?= null
 ): Parcelable
